// src/common/enums/FieldType.ts
/**
 * Enum for field types.
 * @enum {string}
 */
export enum FieldType {
  TEXT = 'text',
  TEXTAREA = 'textarea',
  COLOR = 'color',
  SECURE = 'password',
  NUMBER = 'number',
  EMAIL = 'email',
  DATE = 'date',
  TIME = 'time',
  SELECT = 'select',
  CHECKBOX = 'checkbox',
  SWITCH = 'switch',
  RADIO = 'radio',
  AUTOCOMPLETE = 'autocomplete',
  FILE = 'file',
  DATETIME = 'datetime',
}